+++ CHANGELOG.md
@@ -2,5 +2,5 @@ Please clear changelog after each release.
 Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED.
 -----------------
 - Updated embedded FrozenLib to resolve a crash with Lithium.
-- The Wrench can no longer rotate Chests. ([#1](https://github.com/FrozenBlock/The-Copperier-Age/issues/1))
+- The Wrench can no longer rotate blocks such as Chests, Vaults, Beds, and more. ([#1](https://github.com/FrozenBlock/The-Copperier-Age/issues/1))
   - The blocks that Wrenches cannot rotate is controlled via the new `thecopperierage:wrench_cannot_rotate` block tag.